clean up serializer classes
